backend/requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyJWT==2.8.0
python-dotenv==1.0.0
mysql-connector-python==9.5.0
SQLAlchemy==2.0.44
Werkzeug==3.1.3